+++ mediation/UnityAds/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Unity Ads Unity Mediation Plugin Changelog
 
+#### Next Version
+- Updated the `SetGDPRConsentMetaData(bool)` method from the `UnityAds` class to `SetConsentMetaData(string, bool)`.
+
 #### [Version 3.2.1](https://dl.google.com/googleadmobadssdk/mediation/unity/unity/UnityAdsUnityAdapter-3.2.1.zip)
 - Supports [Unity Ads Android adapter version 4.2.1.1](https://github.com/googleads/googleads-mobile-android-mediation/blob/main/ThirdPartyAdapters/unity/CHANGELOG.md#version-4211).
 - Supports [Unity Ads iOS adapter version 4.2.1.0](https://github.com/googleads/googleads-mobile-ios-mediation/blob/main/adapters/Unity/CHANGELOG.md#version-4210).